+++ src/components/page-wrapper.tsx
@@ -1,15 +1,19 @@
-import { Stack, Title } from '@mantine/core';
+import { Group, Stack, Title } from "@mantine/core";
 
 interface PageWrapperProps {
   name: string;
   children: React.ReactNode;
+  widget?: React.ReactNode;
   height?: number | string;
 }
 
-const PageWrapper: React.FC<PageWrapperProps> = ({ children, name, height }) => {
+const PageWrapper: React.FC<PageWrapperProps> = ({ children, name, height, widget }) => {
   return (
     <Stack spacing="md" h={height}>
-      <Title order={1}>{name}</Title>
+      <Group position="apart">
+        <Title order={1}>{name}</Title>
+        {widget}
+      </Group>
       {children}
     </Stack>
   );