+++ src/features/metrics/components/global-cpu/global-cpu.area-chart.tsx
@@ -1,16 +1,17 @@
-import { useEffect } from 'react';
+import { useEffect } from "react";
 
-import AreaChart, { useAreaChartState } from '@/components/area-chart';
-import Card from '@/components/card';
-import useServerEventsContext from '@/hooks/useServerEventsContext';
-import { useMantineTheme } from '@mantine/core';
+import AreaChart, { useAreaChartState } from "@/components/area-chart";
+import Card from "@/components/card";
+import useGlobalCpuSelectors from "@/features/metrics/stores/global-cpu.store";
+import fromNumberToPercentageString from "@/features/metrics/utils/from-number-to-percentage-string";
+import { useMantineTheme } from "@mantine/core";
 
 // TODO: Remove Luxon and ChartJS
 // TODO: Make timestamp work automatically
 // TODO: fix time
 
 const GlobalCpuAreaChart: React.FC = ({}) => {
-  const { globalCpu } = useServerEventsContext();
+  const metrics = useGlobalCpuSelectors.use.metrics();
   const { other } = useMantineTheme();
   const [chartOptions, setChartOptions] = useAreaChartState({
     title: {
@@ -18,13 +19,15 @@ const GlobalCpuAreaChart: React.FC = ({}) => {
     },
     yAxis: {
       labels: {
-        formatter: (x) => `${x.value}%`,
+        formatter: (x) => `${fromNumberToPercentageString(x.value as number)}`,
       },
       max: 100,
     },
     tooltip: {
       pointFormatter: function () {
-        return `<span style="color:${this.color}">\u25CF</span> ${this.series.name}: <b>${this.y}%</b><br/>`;
+        return `<span style="color:${this.color}">\u25CF</span> ${this.series.name}: <b>${fromNumberToPercentageString(
+          this.y || 0
+        )}</b><br/>`;
       },
     },
   });
@@ -35,12 +38,12 @@ const GlobalCpuAreaChart: React.FC = ({}) => {
         {
           name: "CPU Usage",
           type: "area",
-          data: globalCpu.map((cpu) => [cpu.timestamp, cpu.usage]),
+          data: metrics.map((cpu) => [cpu.timestamp, cpu.usage]),
           color: other.charts.area.globalCpu.color,
         },
       ],
     });
-  }, [globalCpu]);
+  }, [metrics]);
 
   return (
     <Card style={{ height: "410px" }}>